chore(ci) split e2e deployment options (#2950)
It's hard to understand which deployment options only apply to the Kuma
control plane, and which apply to arbitrarry applications. Split the
options in two, and fix up all the relevant option types so that we have
strong typing all the way down, and it's obvious which options apply to
which tasks.

Signed-off-by: James Peach <james.peach@konghq.com>
(cherry picked from commit 96217b56e1852d021db458a4b32ec8c70d7f82d7)

# Conflicts:
#	test/e2e/auth/auth_universal.go